# This makes Chem.SanitizeMol significantly faster when dealing with mo…