@coleb 2016-10-28
Unfortunately, this change was too simplistic. Fragments with 3 atoms can actually be in a ring. In order to skip waters, will need to implement a more rigorous check in the case of 3 atoms to check the atom degrees in the fragment. Closing this pull request for now, will open another in a bit.